requirements.txt
# =========================
# Web Framework
# =========================
fastapi==0.115.5
uvicorn[standard]==0.32.1

# =========================
# Database
# =========================
google-cloud-firestore==2.20.0

# =========================
# Configuration & Validation
# =========================
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# =========================
# Authentication & Security
# =========================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# =========================
# HTTP Client
# =========================
httpx==0.28.1

# =========================
# Cache & Queue
# =========================
redis==5.2.1
celery[redis]==5.4.0

# =========================
# Data Science / BI (CPU Optimized)
# =========================
numpy==2.2.0
pandas==2.2.3
scikit-learn==1.6.0
xgboost==2.1.3
shap==0.46.0

# =========================
# Object Storage
# =========================
minio==7.2.12

# =========================
# Email
# =========================
aiosmtplib==3.0.2

# =========================
# Testing
# =========================
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# =========================
# Linting / Dev Tools
# =========================
ruff==0.8.3

